fixes for this:
the title doesn't look quite centred. This is probably my OC desire for order and structure! Can you move the header a little to the right?
On the "About" page - at the end, we need to take out the reference to modular trainings.
I have seen a couple of references in the Intro course to 6 hours of "continuing professional development" - what does this refer to? If it refers to the fact that it counts as the first course of the modular pathway - it can be taken out.
Why don't we use my radicallyopen.net email address. I haven't been able to add it to my mail  page on my computer, but I can just use the google page and start to build folders there.
On the Events page the Exeter dates are wrong for Part I - should be March 2nd - 6th 2015
On the Contact page there is a phone number after the Lodge address - should this be there - and if so, who is checking it for messages?